+++ src/Bridge/Temporal/Resources/config/services.php
@@ -3,8 +3,10 @@
 declare(strict_types=1);
 
 use Gplanchat\Bridge\Temporal\Command\RunTemporalJournalWorkerCommand;
-use Gplanchat\Bridge\Temporal\Messenger\TemporalJournalTransportFactory;
+use Gplanchat\Bridge\Temporal\Messenger\TemporalTransportFactory;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+
+use function Symfony\Component\DependencyInjection\Loader\Configurator\tagged_iterator;
 
 return static function (ContainerConfigurator $container): void {
     $services = $container->services()
@@ -13,8 +15,9 @@ return static function (ContainerConfigurator $container): void {
         ->autowire()
     ;
 
-    $services->set(TemporalJournalTransportFactory::class)
-        ->tag('messenger.transport_factory')
+    $services->set(TemporalTransportFactory::class)
+        ->args([tagged_iterator('messenger.transport_factory')])
+        ->tag('messenger.transport_factory', ['priority' => -100])
     ;
 
     $services->set(RunTemporalJournalWorkerCommand::class)